+++ R/db-sql.R
@@ -15,6 +15,8 @@
 #' * `sql_random(con)` generates SQL to get a random number which can be used
 #'   to select random rows in `slice_sample()`.
 #'
+#' * `supports_window_clause(con)` does the backend support named windows?
+#'
 #' Tables:
 #'
 #' * `sql_table_analyze(con, table)` generates SQL that "analyzes" the table,
@@ -218,11 +220,11 @@ sql_query_save.DBIConnection <- function(con, sql, name, temporary = TRUE, ...) 
 }
 #' @export
 #' @rdname db-sql
-sql_query_wrap <- function(con, from, name = unique_subquery_name(), ..., lvl = 0) {
+sql_query_wrap <- function(con, from, name = NULL, ..., lvl = 0) {
   UseMethod("sql_query_wrap")
 }
 #' @export
-sql_query_wrap.DBIConnection <- function(con, from, name = unique_subquery_name(), ..., lvl = 0) {
+sql_query_wrap.DBIConnection <- function(con, from, name = NULL, ..., lvl = 0) {
   if (is.ident(from)) {
     setNames(from, name)
   } else if (is.schema(from)) {
@@ -271,6 +273,17 @@ sql_query_rows.DBIConnection <- function(con, sql, ...) {
   build_sql("SELECT COUNT(*) FROM ", from, con = con)
 }
 
+#' @rdname db-sql
+#' @export
+supports_window_clause <- function(con) {
+  UseMethod("supports_window_clause")
+}
+
+#' @export
+supports_window_clause.DBIConnection <- function(con) {
+  FALSE
+}
+
 
 # Query generation --------------------------------------------------------
 
@@ -278,6 +291,7 @@ sql_query_rows.DBIConnection <- function(con, sql, ...) {
 #' @export
 sql_query_select <- function(con, select, from, where = NULL,
                              group_by = NULL, having = NULL,
+                             window = NULL,
                              order_by = NULL,
                              limit = NULL,
                              distinct = FALSE,
@@ -290,6 +304,7 @@ sql_query_select <- function(con, select, from, where = NULL,
 #' @export
 sql_query_select.DBIConnection <- function(con, select, from, where = NULL,
                                group_by = NULL, having = NULL,
+                               window = NULL,
                                order_by = NULL,
                                limit = NULL,
                                distinct = FALSE,
@@ -302,6 +317,7 @@ sql_query_select.DBIConnection <- function(con, select, from, where = NULL,
     where     = sql_clause_where(where),
     group_by  = sql_clause_group_by(group_by),
     having    = sql_clause_having(having),
+    window    = sql_clause_window(window),
     order_by  = sql_clause_order_by(order_by, subquery, limit),
     limit     = sql_clause_limit(con, limit),
     lvl       = lvl
@@ -314,6 +330,7 @@ dbplyr_query_select <- function(con, ...) {
 #' @export
 sql_select.DBIConnection <- function(con, select, from, where = NULL,
                                      group_by = NULL, having = NULL,
+                                     window = NULL,
                                      order_by = NULL,
                                      limit = NULL,
                                      distinct = FALSE,
@@ -324,6 +341,7 @@ sql_select.DBIConnection <- function(con, select, from, where = NULL,
     where = where,
     group_by = group_by,
     having = having,
+    window = window,
     order_by = order_by,
     limit = limit,
     distinct = distinct,